overhauled packagers readers and writers, names and packager abstract base class layout.
