--- uchicagoldrtoolsuite/bit_level/impl/filesystemmaterialsuitestructurepackager.py
@@ -1,57 +1,0 @@
-from ..lib.stagingmaterialsuitepackager import StagingMaterialSuitePackager
-from ..lib.materialsuitestructure import MaterialSuiteStructure
-
-
-__author__ = "Tyler Danstrom"
-__email__ = "tdanstrom@uchicago.edu"
-__company__ = "The University of Chicago Library"
-__copyright__ = "Copyright University of Chicago, 2016"
-__publication__ = ""
-__version__ = "0.0.1dev"
-
-
-class FileSystemMaterialSuiteStructurePackager(
-        StagingMaterialSuitePackager):
-    """
-    Reads a file system MaterialSuite serialization and knows how to package
-    material suites from the contents for inclusion in segment structures
-    """
-    def __init__(self):
-        self.struct_type = 'staging'
-        self.implementation = 'directory'
-
-    def package(self, an_item):
-        msuite = MaterialSuiteStructure(an_item)
-        msuite.original.append(an_item)
-        return msuite
-
-    def get_techmd(self):
-        pass
-
-    def get_presform(self):
-        pass
-
-    def get_premis(self):
-        pass
-
-    def get_struct(self):
-        return self._struct
-
-    def set_struct(self, value):
-        self._struct = value
-
-    def set_type(self, value):
-        self._struct_type = value
-
-    def get_type(self):
-        return self._struct_type
-
-    def set_implementation(self, value):
-        self._implementation = value
-
-    def get_implementation(self):
-        return self._implementation
-
-    struct = property(get_struct, set_struct)
-    struct_type = property(get_type, set_type)
-    implementation = property(get_implementation, set_implementation)
