+++ uchicagoldrtoolsuite/bit_level/lib/segmentpackager.py
@@ -3,8 +3,8 @@ from abc import abstractproperty
 from .abc.packager import Packager
 
 
-__author__ = "Tyler Danstrom"
-__email__ = "tdanstrom@uchicago.edu"
+__author__ = "Brian Balsamo, Tyler Danstrom"
+__email__ = "balsamo@uchicago.edu, tdanstrom@uchicago.edu"
 __company__ = "The University of Chicago Library"
 __copyright__ = "Copyright University of Chicago, 2016"
 __publication__ = ""
@@ -18,11 +18,16 @@ class SegmentPackager(Packager):
     defines the msuite packager they should use to package their contents, as
     well as the interface for setting the segment id
     """
+
+    _msuite_packager = None
+    _id_prefix = None
+    _id_num = None
+
     def set_msuite_packager(self, value):
-        self.msuite_packager = value
+        self._msuite_packager = value
 
     def get_msuite_packager(self):
-        return self.msuite_packager
+        return self._msuite_packager
 
     def set_id_prefix(self, value):
         self._id_prefix = value
@@ -36,7 +41,6 @@ class SegmentPackager(Packager):
     def get_id_num(self):
         return self._id_num
 
-    msuite_packager = abstractproperty(get_msuite_packager,
-                                       set_msuite_packager)
-    id_prefix = abstractproperty(get_id_prefix, set_id_prefix)
-    id_num = abstractproperty(get_id_num, set_id_num)
+    msuite_packager = property(get_msuite_packager, set_msuite_packager)
+    id_prefix = property(get_id_prefix, set_id_prefix)
+    id_num = property(get_id_num, set_id_num)
